secilc: Use size_t for file sizes
The programs secil2conf, secilc2tree, and secilc use uint32_t for
the value of the size of the input files being read. This could
lead to a file larger than 4 GiB being truncated, so use size_t
instead.

Acked-by: Stephen Smalley <stephen.smalley.work@gmail.com>

+++ secilc/secil2tree.c
@@ -71,7 +71,7 @@ int main(int argc, char *argv[])
 	FILE *file = NULL;
 	char *buffer = NULL;
 	struct stat filedata;
-	uint32_t file_size;
+	size_t file_size;
 	char *output = NULL;
 	struct cil_db *db = NULL;
 	int preserve_tunables = 0;
